Refactoring: Move function and everything that belongs to pattern into a seperate component and import that to the dashboard to avoid unnecessary useEffect running
